Time: 38 ms (35.10%) | Memory: 13.1 MB (32.45%) - LeetSync

+++ 843-binary-trees-with-factors/binary-trees-with-factors.cpp
@@ -27,6 +27,6 @@ int MOD=1e9+7;
         for(auto &it:mp){
             res=(res+it.second)%MOD;
         }
-        return res;
+        return (int)res;
     }
 };
